+++ src/core/config.rs
@@ -1,20 +1,33 @@
 use std::fs::{read_to_string, File};
 use std::io::Write;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 use serde::{Deserialize, Serialize};
+
+use anyhow::{anyhow, Context, Result};
 
 #[derive(Serialize, Deserialize)]
 pub struct Config {
     mod_dir: String,
     cache: bool,
+    #[serde(default)]
+    pub game_path: PathBuf,
+    pub nstar_version: Option<String>,
+    #[serde(default)]
+    pub exclude: Vec<String>,
 }
 
 impl Config {
-    pub fn new(dir: String, cache: bool) -> Self {
+    pub fn new(dir: String, cache: bool, game_path: String, nstar_version: Option<String>) -> Self {
         Config {
             mod_dir: dir,
             cache,
+            game_path: PathBuf::from(game_path),
+            nstar_version,
+            exclude: vec![
+                "ns_startup_args.txt".to_string(),
+                "ns_startup_args_dedi.txt".to_string(),
+            ],
         }
     }
 
@@ -35,36 +48,31 @@ impl Config {
     }
 }
 
-pub fn load_config(config_dir: &Path) -> Result<Config, String> {
+pub fn load_config(config_dir: &Path) -> Result<Config> {
     let cfg_path = config_dir.join("config.toml");
     if cfg_path.exists() {
-        let cfg =
-            read_to_string(cfg_path).map_err(|_| ("Unable to read config file".to_string()))?;
-        toml::from_str(&cfg).map_err(|_| ("Unable to parse config".to_string()))
+        let cfg = read_to_string(cfg_path).context("Unable to read config file")?;
+        toml::from_str(&cfg).context("Unable to parse config")
     } else {
-        let mut cfg =
-            File::create(cfg_path).map_err(|_| ("Unable to create config file".to_string()))?;
-        let def = Config::new(String::from("./mods"), true);
-        let parsed = toml::to_string_pretty(&def)
-            .map_err(|_| "Failed to serialize default config".to_string())?;
+        let mut cfg = File::create(cfg_path).context("Unable to create config file")?;
+        let def = Config::new(String::from("./mods"), true, String::new(), None);
+        let parsed = toml::to_string_pretty(&def).context("Failed to serialize default config")?;
         cfg.write_all(parsed.as_bytes())
-            .map_err(|_| "Unable to write config file".to_string())?;
+            .context("Unable to write config file")?;
         Ok(def)
     }
 }
 
-pub fn save_config(config_dir: &Path, config: &Config) -> Result<(), String> {
+pub fn save_config(config_dir: &Path, config: &Config) -> Result<()> {
     let cfg_path = config_dir.join("config.toml");
 
     if cfg_path.exists() {
-        let mut cfg =
-            File::create(&cfg_path).map_err(|_| ("Error opening config file".to_string()))?;
-        let parsed = toml::to_string_pretty(&config)
-            .map_err(|_| ("Error serializing config".to_string()))?;
+        let mut cfg = File::create(&cfg_path).context("Error opening config file")?;
+        let parsed = toml::to_string_pretty(&config).context("Error serializing config")?;
         cfg.write_all(parsed.as_bytes())
-            .map_err(|_| ("Unable to write config file".to_string()))?;
+            .context("Unable to write config file")?;
     } else {
-        return Err("Config file does not exist to write to".to_string());
+        return Err(anyhow!("Config file does not exist to write to"));
     }
     Ok(())
 }