+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Northstar `install`, `init`, and `update` commands
+- File overwrite warning on `update`
+- Overwrite protection for `.cfg` files
+- debug flag for base command
+
+### Chagned
+- Internal error handling
+- Slimmed binary sized a bit 
+
+### Fixed
+- `update` not respecting disabled status
+- `disable` and `enable` not properly modifying all sub mods if the parent mod's name was used
+- `install` potentially using an outdated version from the cache
+
+
+## v2.0.0
+
+### Added
 - `search` command
 - Support "bundle" type mods
 - `enable` and `disable` commands